requirements.txt
# Load's Search — dependencies
# Python 3.8+ recommended

Whoosh==2.7.4
python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
chardet>=5.0.0
